Altered the logging function to use SESSION instead of an array. The prupose is to be able to use logs over file boundries.

+++ delete.php
@@ -1,4 +1,5 @@
 <?php
+session_start();
 /**
  * The purpose of this file is to delete the selected files.
  *
@@ -10,12 +11,18 @@ include("functions.php");
 /**
  * Go through the selected checkboxes and remove the files.
  */
-foreach ($_POST['Subtitle'] as $Subtitle) {
-	if(file_exists($Subtitle)) {
-		if (unlink($Subtitle)) {
-			USMLog("info","Successfully deleted file:" . $Subtitle . "\n");
+if( (count($_POST['Subtitle']) == "0") or (!isset($_POST['Subtitle'])) ) {
+	USMLog("error", "[". __FILE__ ." Line:" . __LINE__ . "] Subtitle POST empty.\n");
+} else {
+	foreach ($_POST['Subtitle'] as $Subtitle) {
+		if(file_exists($Subtitle)) {
+			if (unlink($Subtitle)) {
+				USMLog("info", "[". __FILE__ ." Line:" . __LINE__ . "] Successfully deleted file: " . $Subtitle . "\n");
+			} else {
+				USMLog("error", "[". __FILE__ ." Line:" . __LINE__ . "] Failed deleting file: " . $Subtitle . "\n");
+			}
 		} else {
-			USMLog("error","Failed deleting file:" . $Subtitle . "\n");
+			USMLog("error", "[". __FILE__ ." Line:" . __LINE__ . "] Unable to find the file for deletion: " . $Subtitle . "\n");
 		}
 	}
 }
